Batch Transaction refactored & greeter contract modified
